Added documentation for setting magmom and ldau
