+++ src/aiida_vasp/workchains/v2/vasp.py
@@ -73,8 +73,8 @@ from aiida_vasp.calcs.vasp import VaspCalculation
 from aiida_vasp.common import parameters_validator, warn_deprecated_options
 from aiida_vasp.common.dryrun import get_jobscheme
 from aiida_vasp.data.potcar import PotcarData
-from aiida_vasp.inputset.vaspsets import get_ldau_keys
 from aiida_vasp.protocols import ProtocolMixin, recursive_merge
+from aiida_vasp.utils.ldau import get_ldau_keys
 from aiida_vasp.utils.workchains import compose_exit_code, prepare_process_inputs, site_magnetization_to_magmom
 
 from .mixins import WithBuilderUpdater
